scripts: relative paths and quirks pq rehaul.
More quirks to use relative paths in bitbake (git fetcher and patches).
Instead of applying patches conditionally, now have a generic script
that will apply all series to all relevant repos.

Should a patch not be required, it is easier to comment it in the series
file and re-apply it.

Also, use quilt instead of git directly, as it makes it easier to figure
out if a patch is already applied.

Change-Id: Ic3ea41b3a19dcf2ead1683733d29b2e02d102984
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>